Fix bug in symmetry operator
For our test script, we build a very small symmetry operator. So small that scipy
doesn't return a sparse matrix as a result. This is a problem because we cast the sparse
matrix to a coo_matrix, the form that works best with the proximal operators, so
we first have to check that we have a sparse matrix, then change the type.