Relax symmetry constraint
It was requested that we do not enforce strict symmetry on pixels whose symmetric partner
lies outside of the footprint. So the symmetry operator was modified to allow users to
specify a fill value between 0 and 1, with `fillValue=0` enforcing strict symmetry
(no pixels without a symmetric partner), and `fillValue=1` leaving the pixels unconstrained.