Use even smaller maxiter for testing
Testing shows that most blends converge when the internal deblender iterations are about 50,
and that if we use `maxiter=50` the fluxes we obtain are only about 0.1% worse than using
`maxiter=300`, with a significant improvement in runtime. Optimization later can determine the best
tradeoff between run time and accuracy, but for testing it is to our advantage to use a smaller number
of iterations.