Allow for comparison of multiple deblender methods
As we make modifications to the deblender it will be useful to compare multiple algorithms and operators.
The new simDeblenderTest.ipynb allows a user to pass multiple catalogs, for example one generated using the
monotonicity operator, one using the symmetry operator, one re-apportioning flux based on the NMF deblender
results, and one using the current deblender.