Mask zero pixels by default
`maskPlot` is used to plot an image with some of it's pixels masked out,
making it a nice way to visualize sparse arrays and footprints.
This commit changes it's behavior slightly by masking pixel values
equal to zero by default, as opposed to the previous behavior that
did not apply any mask by default, defeating the purpose of the function.