Prevent model flux from exceeding the data
This commit implements a solution to prevent the model from from
exceeding the data for any pixels. Unfortunately convergence is
very slow and none of the algorithms attempted achieve the desired
functionality, so it has not been cleaned up and is not expected
to be merged, even to my user branch.