Center sources
Modify the `S` (intensity) matrix so that each source sits on the center pixel.
This simplifies the creation of the linear operators, since the same operator works on each source,
and improves the speed of the GLMM algorithm.

In order to do this, this commit also adds two translation operators (`Tx` and `Ty`) that
translate the PSF operator in x and y using a linear interpolation.