Update getFluxPortion to use new models
The `getFluxPortion` method had to be updated to use the new models.
As a side note: the old method of apportioning flux had a bug
in that the SED information wasn't used to create the template weights,
which might explain why the apportioned flux was worse than the
model flux calculations, even before the positions were updated.