Integrate changes from user branch
While the new deblender was being tested a large number of
changes were made to the user branch u/fred/deblender.
The vast majority of those commits contain code that is
no longer necessary or usable in master and it was
thought that merging the entire branch, with dozens of
outdated commits, is undesirable.
Instead the cumulative changes made on the user branch
are contained in this commit, which consists almost
entirely of the work done in DM-11329, which was
already review by pschella in the user branch.